Import cosmos omnibus templates
